+++ ne-routes/neserverroutetest.js
@@ -1,4 +1,10 @@
-var express = require('express');
+if (process.env.NE_AUTO) {
+    var express = require(process.env.NE_AUTO).express
+}
+else {
+    var express = require('express');
+}
+
 var router = express.Router();
 
 var neserverroutetest = function (server){